Final layout for iOS, macOS
